Move NavBar.vue to views directory
