+++ frontend/src/views/Navbar.vue
